Separate methods into files
