+++ src/codel.cpp
@@ -1,0 +1,35 @@
+#include "codel.hpp"
+
+size_t codel_size(const Image& image) {
+  const size_t width = image.get_width();
+  const size_t height = image.get_height();
+  Pixel prev;
+  size_t minimum = gcd(width, height);
+  size_t count = 0;
+  // vertical iterate
+  for (size_t row = 0; row < height; ++row) {
+    for (size_t col = 0; col < width; ++col) {
+      const auto pixel = image[row][col];
+      if (0 < count && prev != pixel) {
+        minimum = gcd(minimum, count);
+        count = 0;
+      }
+      ++count;
+      prev = pixel;
+    }
+  }
+  count = 0;
+  // horizontal iterate
+  for (size_t col = 0; col < width; ++col) {
+    for (size_t row = 0; row < height; ++row) {
+      const auto pixel = image[row][col];
+      if (0 < count && prev != pixel) {
+        minimum = gcd(minimum, count);
+        count = 0;
+      }
+      ++count;
+      prev = pixel;
+    }
+  }
+  return minimum;
+}
