+++ src/codel.hpp
@@ -1,0 +1,8 @@
+#ifndef PIET_I_CODEL_HPP
+#define PIET_I_CODEL_HPP
+
+#include "utils.hpp"
+
+size_t codel_size(const Image& image);
+
+#endif  // PIET_I_CODEL_HPP
